requirements.txt
readabilipy
openai
boto3
transformers
pydub
azure-cognitiveservices-speech
pytz
pytest-playwright
requests_toolbelt
beautifulsoup4
click